+++ .gitignore
@@ -11,3 +11,5 @@ artifacts
 
 .idea
 .DS_Store
+
+typechain-types
