+++ package.json
@@ -10,29 +10,31 @@
   "license": "ISC",
   "dependencies": {
     "@chainlink/contracts": "^0.4.1",
-    "@nomiclabs/hardhat-ethers": "^2.0.5",
-    "@nomiclabs/hardhat-truffle5": "^2.0.5",
-    "@nomiclabs/hardhat-waffle": "^2.0.3",
-    "@nomiclabs/hardhat-web3": "^2.0.0",
     "@openzeppelin/contracts": "4.6.0",
-    "@openzeppelin/contracts-upgradeable": "4.6.0",
-    "@openzeppelin/hardhat-upgrades": "^1.20.1",
-    "@openzeppelin/test-helpers": "^0.5.15",
-    "chai": "^4.3.6",
-    "chai-as-promised": "^7.1.1",
-    "dotenv": "^16.0.0",
-    "ethereum-waffle": "^3.4.4",
-    "ethers": "^5.6.1",
-    "hardhat": "^2.19.5",
-    "hardhat-abi-exporter": "^2.8.0",
-    "hardhat-contract-sizer": "^2.5.1",
-    "hardhat-gas-reporter": "^1.0.8",
-    "hardhat-spdx-license-identifier": "^2.0.3",
-    "hardhat-storage-layout": "^0.1.6",
-    "mocha": "^9.2.2",
-    "solidity-coverage": "^0.7.20"
+    "@openzeppelin/contracts-upgradeable": "4.6.0"
   },
   "devDependencies": {
-    "@nomicfoundation/hardhat-verify": "^2.0.4"
+    "@nomicfoundation/hardhat-chai-matchers": "^2.0.0",
+    "@nomicfoundation/hardhat-ethers": "^3.0.5",
+    "@nomicfoundation/hardhat-network-helpers": "^1.0.0",
+    "@nomicfoundation/hardhat-toolbox": "^4.0.0",
+    "@nomicfoundation/hardhat-verify": "^2.0.4",
+    "@nomiclabs/hardhat-waffle": "^2.0.6",
+    "@openzeppelin/hardhat-upgrades": "^3.0.3",
+    "@typechain/ethers-v6": "^0.5.1",
+    "@typechain/hardhat": "^9.1.0",
+    "@types/chai": "^4.2.0",
+    "@types/mocha": ">=9.1.0",
+    "dotenv": "^16.4.1",
+    "ethers": "^6.10.0",
+    "hardhat": "^2.19.5",
+    "hardhat-abi-exporter": "^2.10.1",
+    "hardhat-contract-sizer": "^2.10.0",
+    "hardhat-gas-reporter": "^1.0.10",
+    "hardhat-spdx-license-identifier": "^2.2.0",
+    "hardhat-storage-layout": "^0.1.7",
+    "mocha": "^10.2.0",
+    "solidity-coverage": "^0.8.6",
+    "typechain": "^8.3.2"
   }
 }
